requirements.txt
asgiref==3.2.7
Django==3.0.7
gunicorn==20.0.4
pytz==2020.1
sqlparse==0.3.1
PyJWT==1.7.1